Fix compiler warnings

+++ libraries/protocol/fee_schedule.cpp
@@ -37,7 +37,7 @@ namespace graphene { namespace protocol {
    fee_schedule fee_schedule::get_default()
    {
       fee_schedule result;
-      for( int i = 0; i < fee_parameters().count(); ++i )
+      for( size_t i = 0; i < fee_parameters().count(); ++i )
       {
          fee_parameters x; x.set_which(i);
          result.parameters.insert(x);
@@ -82,7 +82,7 @@ namespace graphene { namespace protocol {
    {
       auto f = calculate_fee( op, core_exchange_rate );
       auto f_max = f;
-      for( int i=0; i<MAX_FEE_STABILIZATION_ITERATION; i++ )
+      for( size_t i=0; i<MAX_FEE_STABILIZATION_ITERATION; i++ )
       {
          op.visit( set_fee_visitor( f_max ) );
          auto f2 = calculate_fee( op, core_exchange_rate );
